LC-323 - Moved files to identity folder as they are calling the identity service
