+++ test/unit/csrs/service/organisationalUnitServiceTest.ts
@@ -9,7 +9,7 @@ import {OrganisationalUnitService} from '../../../../src/csrs/service/organisati
 import {OrganisationalUnit} from '../../../../src/csrs/model/organisationalUnit'
 import {AgencyTokenHttpService} from '../../../../src/csrs/agencyTokenHttpService'
 import {AgencyToken} from '../../../../src/csrs/model/agencyToken'
-import {AgencyTokenCapacityUsedHttpService} from '../../../../src/csrs/service/agencyTokenCapacityUsedHttpService'
+import {AgencyTokenCapacityUsedHttpService} from '../../../../src/identity/agencyTokenCapacityUsedHttpService'
 
 chai.use(sinonChai)
